requirements.txt
psycopg2-binary==2.9.3
python-dotenv==0.19.2
prettytable==3.12.0